abstract the multi file upload from report form into its own component
